- Finished for validation - Added new error message styles
