+++ client/src/assets/scss/base/_resets.scss
@@ -1,5 +1,6 @@
 * {
-  font-family: "Helvetica Neue", Helvetica, Arial, "Roboto", sans-serif;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Helvetica Neue", "Ubuntu",
+  sans-serif
 }
 
 ::placeholder {
