+++ client/src/assets/scss/blocks/_register-form.scss
@@ -17,7 +17,6 @@
 }
 
 .register-form__input-field {
-  margin: $spacing-xxs 0;
   padding: 0 $spacing-default;
   height: $size-form-field;
   box-sizing: border-box;
@@ -25,10 +24,43 @@
   border: thin solid $border-color-input-field;
   border-radius: $input-field-border-radius;
   transition: box-shadow 0.6s;
+  color: $text-color-input-field;
 
   &:focus {
     outline: none;
     box-shadow: $box-shadow-focus-large;
   }
 }
+
+.register-form__field-wrapper {
+  padding: $spacing-xs 0
+}
+
+.register-form__field-title {
+  color: $color-black-light;
+  font-weight: $font-weight-medium;
+  font-size: $font-size-sm;
+  padding-bottom: $spacing-xs;
+  margin: 0;
+}
+
+.register-form__error-wrapper {
+  padding-top: $spacing-xs;
+}
+
+.register-form__error-icon {
+  width: 12px;
+  height: 12px;
+  padding-right: 8px;
+}
+
+.register-form__error-text {
+  color: $error-color;
+  font-size: $font-size-sm;
+}
+
+.register-form__button {
+  width: 100%;
+}
+
