+++ client/src/assets/scss/styles.scss
@@ -5,6 +5,7 @@
 @import "variables/border-radii";
 @import "variables/box-shadows";
 @import "variables/font-sizes";
+@import "variables/font-weights";
 
 // ** COMPONENT VARIABLES **
 @import "variables/components/button";
