+++ client/src/assets/scss/variables/_font-sizes.scss
@@ -4,6 +4,6 @@ $font-size-h3: 24px;
 $font-size-h4: 20px;
 $font-size-text: 16px;
 $font-size-md: 17px;
-$font-size-sm: 14pxl;
+$font-size-sm: 14px;
 $font-size-xs: 12px;
 $font-size-xxs: 10px;
