+++ client/src/assets/scss/variables/_font-weights.scss
@@ -1,0 +1,1 @@
+$font-weight-medium: 500;
