+++ client/src/components/FormErrorBox/FormErrorBox.types.ts
@@ -1,0 +1,3 @@
+export type FormErrorBoxProps = {
+    error: string,
+}
