+++ client/src/components/FormErrorBox/index.ts
@@ -1,0 +1,2 @@
+export * from './FormErrorBox'
+export * from './FormErrorBox.types'
