+++ client/src/modules/register/index.ts
@@ -1,0 +1,2 @@
+export * from './register'
+export * from './register.types'
