+++ client/src/modules/register/register.types.ts
@@ -1,0 +1,5 @@
+export type UserErrors = {
+    username?: string,
+    email?: string,
+    password?: string,
+}
