+++ server/src/typeDefs/user.type-def.ts
@@ -1,10 +1,6 @@
 import { gql } from 'apollo-server'
 
 export const userType = gql`
-    type Query {
-        getAllUsers: [User]!
-    }
-
     type User {
         id: ID!
         email: String!
@@ -12,8 +8,12 @@ export const userType = gql`
         password: String!
     }
 
+    type Query {
+        getAllUsers: [User]!
+    }
 
     type Mutation {
         createUser(email: String!, username: String!, password: String!): User
     }
+
 `
